Start experimental ARM64 Android build

+++ include/types.h
@@ -36,11 +36,11 @@
 
 typedef signed char s8;
 typedef signed short s16;
-typedef signed long s32;
+typedef signed int s32;
 typedef signed long long s64;
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 #ifndef TARGET_PC
 #ifndef _SIZE_T_DEF
 #define _SIZE_T_DEF
